// libs/shared/modules/src/data/models/domain/policy.ts
/**
 * Anecdotes API
 * Anecdotes API's
 *
 * OpenAPI spec version: 0.0.3
 * Contact: apiteam@anecdotes.ai
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CombinedEvidenceInstance } from './combinedEvidenceInstance';
import { PolicySettings } from './policySettings';

export interface Policy {
  /**
   * The policy unique ID
   */
  readonly policy_id?: string;
  /**
   * The policy name
   */
  policy_name?: string;
  /**
   * The policy description
   */
  policy_description?: string;
  /**
   * Has template url
   */
  policy_has_template?: boolean;
  /**
   * policy group
   */
  policy_group?: string;
  /**
   * The user who last added/edited this policy
   */
  policy_edited_by?: string;
  /**
   * The time this policy was last added/edited
   */
  policy_last_edit_time?: string;
  /**
   * Whether this policy is applicable for the customer or not
   */
  policy_is_applicable?: boolean;
  /**
   * Whether this policy was created or edit by the user
   */
  policy_is_custom?: boolean;
  /**
   * The related evidence
   */
  policy_related_evidence?: string;
  /**
   * Boolean to determine whether a note exists for this policy
   */
  note_exists?: boolean;
  policy_settings?: PolicySettings;
  evidence?: CombinedEvidenceInstance;
  policy_related_frameworks_names?: { [framework_name: string]: string[]; };
  notify_me_scheduledtask_id?: string;
  notify_approvers_scheduledtask_id?: string;
  snapshot_id?: string;
  policy_type?: string;
}

export function isPolicy(object: any): object is Policy {
  return 'policy_id' in object;
}