// libs/shared/modules/src/data/models/domain/scheduleSettings.ts
/**
 * Anecdotes API
 * Anecdotes API's
 *
 * OpenAPI spec version: 0.0.3
 * Contact: apiteam@anecdotes.ai
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ScheduleSettings {
    approval_frequency?: ApprovalFrequencyEnum;
    start_from?: Date;
    notify_me?: NotifyMeEnum;
    notify_approvers?: NotifyApproversEnum;
}
export type ApprovalFrequencyEnum = 'yearly' | 'quarterly' | 'monthly';
export const ApprovalFrequencyEnum = {
    Yearly: 'yearly' as ApprovalFrequencyEnum,
    Quarterly: 'quarterly' as ApprovalFrequencyEnum,
    Monthly: 'monthly' as ApprovalFrequencyEnum
};
export type NotifyMeEnum = 'never' | 'day' | 'week' | 'month';
export const NotifyMeEnum = {
    Never: 'never' as NotifyMeEnum,
    Day: 'day' as NotifyMeEnum,
    Week: 'week' as NotifyMeEnum,
    Month: 'month' as NotifyMeEnum
};
export type NotifyApproversEnum = 'never' | 'day' | 'week' | 'month';
export const NotifyApproversEnum = {
    Never: 'never' as NotifyApproversEnum,
    Day: 'day' as NotifyApproversEnum,
    Week: 'week' as NotifyApproversEnum,
    Month: 'month' as NotifyApproversEnum
};